Releasing version 2.6.3
