+++ requirements.txt
@@ -16,7 +16,7 @@ Jinja2==2.10.1
 jmespath==0.9.3
 ndg-httpsclient==0.4.2
 mock==2.0.0
-oci==2.3.2
+oci==2.3.3
 packaging==16.8
 pluggy==0.4.0
 py==1.4.33
@@ -38,4 +38,4 @@ tox==2.9.1
 vcrpy==1.13.0
 virtualenv==15.1.0
 pytest-xdist==1.22.2
-PyYAML==3.13
+PyYAML==5.1.2
