+++ services/core/examples_and_test_scripts/cluster_network_example_files/instance_config_details_template.json
@@ -1,0 +1,16 @@
+{
+  "instanceType": "compute",
+  "launchDetails": {
+    "compartmentId": "__COMPARTMENT_ID__",
+    "shape": "BM.HPC2.36",
+    "sourceDetails":
+      {
+        "imageId": "__IMAGE_ID__",
+        "sourceType": "image",
+        "bootVolumeSizeInGBs": 75
+      },
+    "createVnicDetails": {
+      "assignPublicIp": true
+    }
+  }
+}
