+++ services/core/examples_and_test_scripts/cluster_network_example_files/instance_pools_input_template.json
@@ -1,0 +1,6 @@
+[
+  {
+    "instanceConfigurationId": "__INSTANCE_CONFIG_ID__",
+    "size": "1"
+  }
+]
