+++ services/core/examples_and_test_scripts/cluster_network_example_files/placement_details_template.json
@@ -1,0 +1,4 @@
+{
+  "availabilityDomain": "__AD__",
+  "primarySubnetId": "__SUBNET_ID__"
+}
