+++ src/oci_cli/__init__.py
@@ -63,9 +63,10 @@ from . import cli_session  # noqa: F401,E402
 from . import cli_setup_bootstrap  # noqa: F401,E402
 from . import cli_util  # noqa: F401,E402
 from . import cli_exceptions  # noqa: F401,E402
-import services.object_storage.src.oci_cli_object_storage.object_storage_transfer_manager  # noqa: F401,E402
 from . import json_skeleton_utils  # noqa: F401,E402
 from . import string_utils  # noqa: F401,E402
 from . import help_text_producer  # noqa: F401,E402
 from oci import config  # noqa: F401,E402
 from .version import __version__  # noqa: F401,E402
+
+final_command_processor.process()
